refactor: 로또 구입 관련 검증 기능을 LottoMachine 클래스로 이동
